% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LBL.ex2}
\alias{LBL.ex2}
\title{This dataset consists of affection status, complex sampling weights, a binary stratifying
variable, a binary environmental covariate, and SNP data.}
\format{A data frame with variables: \code{affected}, \code{wt},
 \code{stra}, \code{cov},
 \code{M1.1}, \code{M1.2}, \code{M2.1}, \code{M2.2}, \code{M3.1},
 \code{M3.2}, \code{M4.1}, \code{M4.2}, \code{M5.1}, \code{M5.2}.}
\usage{
LBL.ex2
}
\description{
This dataset consists of affection status, complex sampling weights, a binary stratifying
variable, a binary environmental covariate, and SNP data.
}
\keyword{datasets}
