% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LBL.ex3}
\alias{LBL.ex3}
\title{This dataset consists of two correlated affection statuses, no environmental covariate,
and SNP data.}
\format{A data frame with variables: \code{Y1}, \code{Y2},
  \code{x1}, \code{x2}, \code{x3}, \code{x4}, \code{x5},
  \code{x6}, \code{x7}, \code{x8}, \code{x9}, \code{x10}.}
\usage{
LBL.ex3
}
\description{
This dataset consists of two correlated affection statuses, no environmental covariate,
and SNP data.
}
\keyword{datasets}
