\name{LBLGXE-package}
\alias{LBLGXE-package}
\alias{LBLGXE}
\docType{package}
\title{
Logistic Bayesian Lasso for Rare (or Common) Haplotype Association
}
\description{
The main function of this package is LBL. For details, see ?LBL.
}
\details{
\tabular{ll}{
Package: \tab LBLGXE\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2019-05-17\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
Currently available functions: LBL. Type ?LBL for more details.
}
\author{
Xiaochen Yuan, Yuan Zhang, Shuang Xia, Swati Biswas, and Shili Lin

Maintainer: Xiaochen Yuan<xxy142030@utdallas.edu>
}
\references{
Yuan X and Biswas S. Bivariate Logistic Bayesian LASSO for Detecting Rare Haplotype Association with Two Correlated Phenotypes. Under review.

Zhang Y, Hofmann J, Purdue M, Lin S, and Biswas S.
Logistic Bayesian LASSO for Genetic Association Analysis of Data from Complex Sampling Designs. Journal of Human Genetics, 62:819-829.

Zhang Y, Lin S, and Biswas S.
Detecting Rare and common Haplotype-Environment Interaction under Uncertainty of Gene-Environment Independence Assumption. Biometrics, 73:344-355.

Zhang, Y. and Biswas, S (2015). An Improved Version of Logistic Bayesian LASSO for Detecting Rare Haplotype-Environment Interactions With Application to Lung Cancer, Cancer Informatics, 14(S2): 11-16.

Biswas S, Xia S and Lin S (2014). Detecting Rare Haplotype-Environment Interaction with Logistic Bayesian LASSO. Genetic Epidemiology, 38: 31-41.

Biswas S and Lin S (2012). Logistic Bayesian LASSO for Identifying Association with Rare Haplotypes and Application to Age-related Macular Degeneration. Biometrics, 68(2): 587-97.

Burkett K, Graham J and McNeney B (2006). hapassoc: Software for Likelihood Inference of Trait Associations with SNP Haplotypes and Other Attributes. Journal of Statistical Software, 16(2): 1-19.

}
\keyword{ package }
\examples{
#see ?LBL
}

