% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LBL.ex3}
\alias{LBL.ex3}
\title{This dataset consists of two correlated affection statuses, no environmental covariate,
and SNP data.}
\format{A data frame with variables: \code{Y1}, \code{Y2},
  \code{M1.1}, \code{M1.2}, \code{M2.1}, \code{M2.2}, \code{M3.1},
  \code{M3.2}, \code{M4.1}, \code{M4.2}, \code{M5.1}, \code{M5.2}.}
\usage{
LBL.ex3
}
\description{
This dataset consists of two correlated affection statuses, no environmental covariate,
and SNP data.
}
\keyword{datasets}
